% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_render.R
\name{render}
\alias{render}
\title{Rmarkdown render replacement}
\usage{
render(input, encoding)
}
\description{
The default Rmarkdown render in Rstudio doesn't play nice with our datakit
project structure. This replacement function does a better job of separating
the source .Rmd file from the output .html file
}
